Support dictionaries with different entries below same parent
Add checkbox for empty translation in order to keep them empty
Default behaviour for empty translations is still removing the
underlying entry node.
Expose keys in a way that special characters are visible.
Escape newlines in keys as they are used in data attributes in HTML, get
rid of hidden form field "key" in Update dialog to prevent issues with
newlines.

This closes #207 #172 #208